% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butterfly}
\alias{butterfly}
\title{Frequencies of butterflies}
\format{A data frame with 25 rows and 2 variables:
\describe{
  \item{index}{An index variable j}
  \item{frequency}{The number of butterflies observed j times}
  ...
}}
\source{
Fisher, R.A., Corbet A.S., & Williams, C.B.. 
"The relation between the number of species and 
the number of individuals in a random sample of an animal population." 
The Journal of Animal Ecology (1943): 42-58.
}
\usage{
butterfly
}
\description{
A dataset containing a frequency count table of butterflies in 
British Malaya in the 1940's
}
\keyword{datasets}
