% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butterfly_catchall}
\alias{butterfly_catchall}
\title{The CatchAll output of butterfly}
\format{A data frame with 248 rows rows and 20 variables:}
\source{
Fisher, R.A., Corbet A.S., & Williams, C.B.. 
"The relation between the number of species and 
the number of individuals in a random sample of an animal population." 
The Journal of Animal Ecology (1943): 42-58.

Bunge, John, et al. "Estimating population diversity with 
CatchAll." Bioinformatics 28.7 (2012): 1045-1047.
}
\usage{
butterfly_catchall
}
\description{
The estimated species richness of butterflies
}
\keyword{datasets}
