% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_all_main.R
\name{catch_all}
\alias{catch_all}
\title{CatchAll for estimating species richness and model selection}
\usage{
catch_all(input_data, ...)
}
\arguments{
\item{input_data}{An input type that can be processed by \code{convert()}}

\item{...}{Other arguments for \code{select_best_models}}
}
\description{
This function implements a series of estimator and selects several of them for estimating species richness
}
\examples{
library(breakaway)
data(apples)
catch_all(apples)

}
