% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchAll.R
\name{CatchAll}
\alias{CatchAll}
\title{Species richness estimation with CatchAll}
\usage{
CatchAll(frequency_table)
}
\arguments{
\item{frequency_table}{The sample frequency count table for the population of
interest. The first row must correspond to the singletons. Acceptable
formats include a matrix, data frame, or file path (csv or txt). The
standard frequency count table format is used: two columns, the first of
which contains the frequency of interest (eg. 1 for singletons, species
observed once, 2 for doubletons, species observed twice, etc.) and the
second of which contains the number of species observed this many times.
Frequencies (first column) should be ordered least to greatest.}
}
\value{
All models and their species richness estimates (beta version)
}
\description{
This function implements the CatchAll species richness estimate in R. The packages currently in beta mode!
}
\author{
Amy Willis

Teresa Zhan
}
