% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{CheckInput}
\alias{CheckInput}
\title{CheckInput}
\usage{
CheckInput(frequency_count)
}
\arguments{
\item{frequency_count}{The frequency count table to check}
}
\value{
The checked and fixed frequency count table
}
\description{
Check and fix the formatting of a frequency count table
}
