% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_models.R
\name{GOF}
\alias{GOF}
\title{Goodness-of-fit statistic}
\usage{
GOF(O_c, E_c, param, bin.tol = 0)
}
\arguments{
\item{O_c}{Observed counts in the bins;}

\item{E_c}{Expected counts in the bins;}

\item{bin.tol}{If the expected count in one bin is less than bin.tol, we merge this and the next bin together. We continue this process until the counts in every bin is larger than bin.tol.}
}
\value{
a list containing the Goodness-of-fit statistic (GOF) and degree of freedom (df).
}
\description{
Given the observed and expected counts in different bins, this function returns the goodness-of-fit statistic
}
