% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{PoissonModel}
\alias{PoissonModel}
\title{PoissonModel}
\usage{
PoissonModel(frequency_count, cutoff = NULL)
}
\description{
A model to estimate the number of missing taxa under a Poisson Model
}
\examples{
library(breakaway)
data(apples)
PoissonModel(apples, 4)

}
