% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_mixed_exp.R
\name{three_mixed_exp_lld}
\alias{three_mixed_exp_lld}
\title{Internal functions for estimating species richness with a three-component geomic mixture model}
\usage{
three_mixed_exp_lld(ii, ff, t1, t2, t3, u1, u2, full = F)
}
\description{
Internal functions for estimating species richness with a three-component geomic mixture model
}
