% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{PoissonModel}
\alias{PoissonModel}
\title{PoissonModel}
\usage{
PoissonModel(frequency_count, cutoff = NULL)
}
\arguments{
\item{frequency_count}{A frequency count table}

\item{cutoff}{The largest frequency to use for predicting f0}
}
\description{
A model to estimate the number of missing taxa under a Poisson Model
}
\examples{
data(butterfly)
PoissonModel(butterfly, 4)

}
